# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AvailabilityResponse": {
        "properties": {
          "date": {
            "description": "The requested date, interpreted in the business timezone.",
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "service_id": {
            "title": "Service Id",
            "type": "string"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/SlotResponse"
            },
            "title": "Slots",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "service_id",
          "date",
          "timezone",
          "slots"
        ],
        "title": "AvailabilityResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorFieldDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorFieldDetail": {
        "properties": {
          "issue": {
            "description": "Machine-readable validation issue type.",
            "title": "Issue",
            "type": "string"
          },
          "loc": {
            "description": "Location of the invalid field.",
            "items": {
              "type": "string"
            },
            "title": "Loc",
            "type": "array"
          }
        },
        "required": [
          "loc",
          "issue"
        ],
        "title": "ErrorFieldDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "SlotResponse": {
        "properties": {
          "end": {
            "description": "Slot end, UTC.",
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "start": {
            "description": "Slot start, UTC.",
            "format": "date-time",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "title": "SlotResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Fictional business data and deterministic availability. Demo data only.",
    "title": "Voice Agent Sandbox API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/services/{service_id}/availability": {
      "get": {
        "operationId": "read_availability_v1_services__service_id__availability_get",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "string"
            }
          },
          {
            "description": "Local date in the business timezone (YYYY-MM-DD).",
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "description": "Local date in the business timezone (YYYY-MM-DD).",
              "format": "date",
              "title": "Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown service."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or date outside the window."
          }
        },
        "summary": "Read Availability",
        "tags": [
          "availability"
        ]
      }
    }
  }
}